+++ packages/terra-core-docs/src/terra-dev-site/test/section-header/OnClickSectionHeader.test.jsx
@@ -3,7 +3,7 @@ import SectionHeader from 'terra-section-header';
 
 const OnClickSectionHeader = () => (
   <SectionHeader
-    title="OnClick Section Header"
+    text="OnClick Section Header"
     // eslint-disable-next-line no-alert
     onClick={() => { window.alert('The accordion has been clicked!'); }}
     className="onClickHeader"